requirements.txt
pandas>=2.2
pyarrow>=16.0
boto3>=1.34
python-dotenv>=1.0
PyYAML>=6.0
psycopg2-binary>=2.9   # (PostgreSQL 쓰실 경우)
